+++ apps/site/lib/site_web/controllers/schedule/schedule_api.ex
@@ -1,6 +1,11 @@
 defmodule SiteWeb.ScheduleController.ScheduleApi do
   use SiteWeb, :controller
 
+  alias Routes.Route
+  alias Fares.Format
+  alias Site.{BaseFare}
+  import SiteWeb.ViewHelpers, only: [cms_static_page_path: 2]
+
   def show(conn, %{"id" => route_id, "date" => date, "direction_id" => direction_id}) do
     {:ok, date} = Date.from_iso8601(date)
     schedule_data = get_schedules(route_id, date, direction_id)
@@ -10,18 +15,85 @@ defmodule SiteWeb.ScheduleController.ScheduleApi do
 
   @spec get_schedules(binary, any, any) :: %{by_trip: map, trip_order: [any]}
   def get_schedules(route_id, date, direction_id) do
-    schedules =
-      Enum.map(
-        Schedules.Repo.by_route_ids([route_id], date: date, direction_id: direction_id),
-        &Map.update!(&1, :route, fn route -> Routes.Route.to_json_safe(route) end)
-      )
+    services =
+      [route_id]
+      |> Schedules.Repo.by_route_ids(date: date, direction_id: direction_id)
+      |> Enum.map(&Map.update!(&1, :route, fn route -> Route.to_json_safe(route) end))
 
-    schedules_by_trip =
-      schedules
-      |> Enum.map(&Map.update!(&1, :time, fn time -> Timex.format!(time, "{0h12}:{m} {AM}") end))
+    ordered_trips = services |> Enum.sort_by(& &1.time) |> Enum.map(& &1.trip.id) |> Enum.uniq()
+
+    services_by_trip =
+      services
       |> Enum.group_by(& &1.trip.id)
 
-    ordered_trips = schedules |> Enum.sort_by(& &1.time) |> Enum.map(& &1.trip.id) |> Enum.uniq()
-    %{by_trip: schedules_by_trip, trip_order: ordered_trips}
+    services_by_trip_with_fare =
+      services_by_trip
+      |> Stream.map(fn {trip_id, service} -> {trip_id, fares_for_service(service)} end)
+      |> Stream.map(fn {trip_id, service} -> {trip_id, duration_for_service(service)} end)
+      |> Stream.map(fn {trip_id, service} -> {trip_id, formatted_time(service)} end)
+      |> Enum.into(%{})
+
+    %{by_trip: services_by_trip_with_fare, trip_order: ordered_trips}
+  end
+
+  def fares_for_service(schedules) do
+    origin = List.first(schedules)
+
+    schedules
+    |> Enum.map(
+      &Map.merge(
+        &1,
+        fares_for_service(origin.route, origin.stop.id, &1.stop.id)
+      )
+    )
+  end
+
+  def duration_for_service(schedules) do
+    first = List.first(schedules).time
+    last = List.last(schedules).time
+    %{schedules: schedules, duration: Timex.diff(last, first, :minutes)}
+  end
+
+  def formatted_time(%{schedules: schedules, duration: duration}) do
+    time_formatted_schedules =
+      schedules
+      |> Enum.map(&Map.update!(&1, :time, fn time -> Timex.format!(time, "{0h12}:{m} {AM}") end))
+
+    %{schedules: time_formatted_schedules, duration: duration}
+  end
+
+  @spec fares_for_service(map, String.t(), String.t()) :: map
+  def fares_for_service(route, origin, destination) do
+    %{
+      price: route |> BaseFare.base_fare(origin, destination) |> Format.price(),
+      fare_link:
+        fare_link(
+          Route.type_atom(route.type),
+          origin,
+          destination
+        )
+    }
+  end
+
+  def fare_link(:bus, _origin, _destination) do
+    cms_static_page_path(SiteWeb.Endpoint, "/fares/bus-fares")
+  end
+
+  def fare_link(:subway, _origin, _destination) do
+    cms_static_page_path(SiteWeb.Endpoint, "/fares/subway-fares")
+  end
+
+  def fare_link(:commuter_rail, origin, destination) do
+    fare_path(SiteWeb.Endpoint, :show, :commuter_rail, %{
+      origin: origin,
+      destination: destination
+    })
+  end
+
+  def fare_link(:ferry, origin, destination) do
+    fare_path(SiteWeb.Endpoint, :show, :ferry, %{
+      origin: origin,
+      destination: destination
+    })
   end
 end
